requirements.txt
# requirements.txt
flask==3.0.0
opencv-python==4.9.0.80
ultralytics==8.0.230
torch==2.1.2
torchvision==0.16.2
numpy==1.24.3
Pillow==10.1.0
pyyaml==6.0.1